Move config dir setup to separate function
Nicer to work with than macro kerfuffle.
